refactor: 행거 기울어지는 애니메이션 훅으로 분리
