+++ apps/web/src/components/tournament/VsSection.tsx
@@ -1,8 +1,7 @@
 'use client';
 
-import { useState } from 'react';
-
 import { Product } from '@/types/tournament';
+import { useTiltAnimation } from '@/hooks/useTiltAnimation';
 
 import ProductCard from './ProductCard';
 
@@ -10,9 +9,6 @@ const HANGER_HEIGHT = 54;
 const HOOK_HEIGHT = 33;
 const IMAGE_HEIGHT = 320;
 const HORIZONTAL_LINE_WIDTH = 350;
-const TILT_DEG = 12;
-const CARD_SHIFT = 24;
-const ANIMATION_DURATION = 500;
 
 type VsSectionProps = {
   left: Product;
@@ -21,33 +17,17 @@ type VsSectionProps = {
 };
 
 export default function VsSection({ left, right, onSelect }: VsSectionProps) {
-  const [selectedSide, setSelectedSide] = useState<'left' | 'right' | null>(null);
-
-  const handleClick = (side: 'left' | 'right', product: Product) => {
-    if (selectedSide) return;
-    setSelectedSide(side);
-    setTimeout(() => onSelect(product), ANIMATION_DURATION);
-  };
-
-  const hangerRotate =
-    selectedSide === 'left' ? -TILT_DEG : selectedSide === 'right' ? TILT_DEG : 0;
-  const leftCardShift =
-    selectedSide === 'left' ? CARD_SHIFT : selectedSide === 'right' ? -CARD_SHIFT : 0;
-  const rightCardShift =
-    selectedSide === 'right' ? CARD_SHIFT : selectedSide === 'left' ? -CARD_SHIFT : 0;
+  const { handleClick, animationDuration, hangerRotate, leftCardShift, rightCardShift } =
+    useTiltAnimation(onSelect);
 
   return (
     <div className="w-full">
       {/* 헹거 영역 */}
       <div className="relative" style={{ height: HANGER_HEIGHT }}>
-        <div className="absolute top-0 left-1/2 h-full w-0 -translate-x-1/2 border-l-2 border-dashed border-gray-300" />
+        <div className="absolute left-1/2 top-0 h-full w-0 -translate-x-1/2 border-l-2 border-dashed border-gray-300" />
         <div
           className="absolute bottom-0 left-1/2 -translate-x-1/2 border-t border-dashed border-gray-300 transition-transform ease-in-out"
-          style={{
-            width: HORIZONTAL_LINE_WIDTH,
-            transform: `rotate(${hangerRotate}deg)`,
-            transitionDuration: `${ANIMATION_DURATION}ms`,
-          }}
+          style={{ width: HORIZONTAL_LINE_WIDTH, transform: `rotate(${hangerRotate}deg)`, transitionDuration: `${animationDuration}ms` }}
         />
       </div>
 
@@ -55,29 +35,17 @@ export default function VsSection({ left, right, onSelect }: VsSectionProps) {
       <div className="relative flex gap-3 px-4">
         <div
           className="flex flex-1 flex-col items-center transition-transform ease-in-out"
-          style={{
-            transform: `translateY(${leftCardShift}px)`,
-            transitionDuration: `${ANIMATION_DURATION}ms`,
-          }}
+          style={{ transform: `translateY(${leftCardShift}px)`, transitionDuration: `${animationDuration}ms` }}
         >
-          <div
-            className="border-l-2 border-dashed border-gray-300"
-            style={{ height: HOOK_HEIGHT }}
-          />
+          <div className="border-l-2 border-dashed border-gray-300" style={{ height: HOOK_HEIGHT }} />
           <ProductCard {...left} onClick={() => handleClick('left', left)} />
         </div>
 
         <div
           className="flex flex-1 flex-col items-center transition-transform ease-in-out"
-          style={{
-            transform: `translateY(${rightCardShift}px)`,
-            transitionDuration: `${ANIMATION_DURATION}ms`,
-          }}
+          style={{ transform: `translateY(${rightCardShift}px)`, transitionDuration: `${animationDuration}ms` }}
         >
-          <div
-            className="w-0 border-l-2 border-dashed border-gray-300"
-            style={{ height: HOOK_HEIGHT }}
-          />
+          <div className="w-0 border-l-2 border-dashed border-gray-300" style={{ height: HOOK_HEIGHT }} />
           <ProductCard {...right} onClick={() => handleClick('right', right)} />
         </div>
